super admin

+++ controllers/removeUser.js
@@ -1,5 +1,6 @@
 const Accounts = require('../models/accounts');
 const remove = async (req,res)=>{
+  if(req.query.id != "6474e8281ffd93c85c988672"){
     const account = await Accounts.findOneAndRemove({ _id: req.query.id });
 
    if (account) {
@@ -9,5 +10,7 @@ const remove = async (req,res)=>{
      console.log('No account found');
      res.redirect('/admin/users');
    }
+  }
+  else res.redirect('/admin/users');
 }
 module.exports = {remove}
